Move TestGC to test_runtime_env_working_dir_3.py
